add create activitie

+++ database/migrations/2024_09_10_075207_create_activities_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
 
             $table->text('notes')->nullable();
 
-            $table->unsignedTinyInteger('satisfaction')->nullable()->default(null);
+            $table->integer('satisfaction')->nullable()->default(null);
 
             $table->timestamps();
         });
